pyproject.toml
[project]
name = "raspberry-pi-tester"
version = "0.1.0"
description = "Hardware test suite for Raspberry Pi 4B"
requires-python = ">=3.11"
dependencies = [
    "RPi.GPIO>=0.7.1",
    "smbus2>=0.4.3",
    "picamera2>=0.3.12",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
